# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Amazing Studio - Wedding Studio Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dresses:
    get:
      operationId: listDresses
      tags: [dresses]
      parameters:
        - name: available
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dress"
    post:
      operationId: createDress
      tags: [dresses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDressRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dress"
  /dresses/{id}:
    get:
      operationId: getDress
      tags: [dresses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dress"
    put:
      operationId: updateDress
      tags: [dresses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDressRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dress"
    delete:
      operationId: deleteDress
      tags: [dresses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Dress:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        color:
          type: string
        size:
          type: string
        style:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        rentalPrice:
          type: number
        depositRequired:
          type: number
        isAvailable:
          type: boolean
        condition:
          type: string
          enum: [excellent, good, fair, needs_repair]
        notes:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, code, name, color, size, rentalPrice, depositRequired, isAvailable, condition, createdAt]
    CreateDressRequest:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        color:
          type: string
        size:
          type: string
        style:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        rentalPrice:
          type: number
        depositRequired:
          type: number
        condition:
          type: string
          enum: [excellent, good, fair, needs_repair]
        notes:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required: [code, name, color, size, rentalPrice, depositRequired, condition]
...